// packages/@isomorphic-typescript/ts-monorepo~/source/common/types/monorepo-package.ts
import { MergedPackageConfig } from "./merged-config";

export interface MonorepoPackage {
    relativePath: string;
    name: string;
    version: string;
    relationships: {
        dependsOn: Record<string, MonorepoPackage>;
        dependencyOf: Record<string, MonorepoPackage>;
    };
    config: MergedPackageConfig;
}